requirements.txt
# Django and Core Dependencies
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0

# Authentication & Security
PyJWT==2.10.1
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0

# Configuration Management
python-decouple==3.8

# HTTP Requests
requests==2.32.5

# Google API Client Libraries
google-api-python-client==2.149.0
google-api-core==2.19.2

# AI/ML Libraries
openai==1.54.3

# Database
psycopg2-binary==2.9.10
